+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 #### 1.8 (UNRELEASED)
 
 New features and improvements:
-* ...
+* Added `lswap` command to swap the content of two layers (#300)
+* Added `lreverse` command to reverse the order of paths within a layer (#300)
 
 Bug fixes:
 * ...